fix: 파이프라인 복구 + README/문서 현행화
- selector: 카테고리 배치 JSON truncation으로 전 기사 0점 처리되던 조용한
  실패 수정 (후보 사전제한 + 청크 분할 + 객체 단위 복구 파싱 + json_mode)
- config: ~/.hermes/.skills.env 폴백 로드로 NOTION_API_KEY 미설정 오류 해결
- collector: arxiv.Client로 429 완화, Hada.io 이중 수집 제거
- auto_publish_approved: 하드코딩 경로 제거(config 사용), 동일 DB 정합,
  발행 상태/건수 환경변수화(AUTO_PUBLISH_STATUS=검토 완료, AUTO_PUBLISH_MAX=5)
- README: Ollama Gemma4·Hermes cron·Hugo·실제 경로/스크립트로 전면 현행화
- .gitignore 추가, 추적되던 __pycache__/news.db/.omc 제거

+++ scripts/models.py
@@ -8,7 +8,7 @@ from typing import List, Optional
 
 class EvaluationItem(BaseModel):
     id: int = Field(..., description="기사 ID")
-    score: int = Field(..., ge=1, le=10, description="평가 점수 (1-10)")
+    score: float = Field(..., ge=1, le=10, description="평가 점수 (1-10)")
     reason: str = Field(..., description="선별 이유")
 
 class EvaluationResponse(BaseModel):
